refactor: 클린 아키텍처를 적용하여 전체 코드 수정
