--- back/express/validators/index.js
@@ -1,5 +1,0 @@
-const mapState = require('./mapState');
-
-module.exports={
-    mapState
-}
