+++ back/src/controllers/displayDealsController.js
@@ -1,0 +1,25 @@
+const LocationService = require('../application/service/location');
+const LocationFormService = require('../application/service/locationForm');
+
+const LocationRepository = require('../infrastructure/database/repositories/LocationRepositoryMysql');
+const LocationFormRepository = require('../infrastructure/database/repositories/LocationFormRepositoryMysql')
+
+exports.basic=async (options)=>{
+
+    const locationService= new LocationService(new LocationRepository());
+    const locationFormService= new LocationFormService(new LocationFormRepository());
+    const coordinate={
+        min_x:options.body.bounds.min.x,
+        min_y:options.body.bounds.min.y,
+        max_x:options.body.bounds.max.x,
+        max_y:options.body.bounds.max.y,
+    }
+
+    try{
+        let locationList = await locationService.findIncluedArea(coordinate);
+        return await locationFormService.findRecentlyDealsDistinct(coordinate,locationList)
+        
+    }catch(error){
+        console.log(error)
+    }
+}
