+++ back/src/infrastructure/database/orm/sequelize/models/location.js
